Added a JSON-unpacking function which acts according to the event type.
Right now, the function generates default descriptions according to
what the event type received is. In the future, we may want to
change the JSON schema to allow for the user scanning his phone
to add his/her own description.

Also, modified the addRaspberry function so that it now checks
whether the given UUID is already in use, and adds the new raspberry
if it's not in use. A boolean is returned (True if the raspberry
was succesfully added into the database).