Added generateJSON and fixed Event representation.
generateJSON takes in a user ID as input, finds all unsent events
for that user, and packs them into a JSON which has a "newEvents"
object containing an array of the unsent event details. 'sent' is
updated to True after adding the event into eventList.

When using SQLAlchemy for queries, events now show 'sent' as well.