Renamed unpackJSON and generateJSON and split them into several functions.
This was to make the function more flexible and easier to read.
postJSON (renamed from unpackJSON) was split into piPosts,
phonePosts and phoneGets.

piPosts deals with POST requests from the raspberry pi and returns
a boolean whether an event was added to the table or not.

phonePosts deals with POST requests from phones. If this is a
REGISTER request, it returns the new userid if successful and False
if not. For ADDPI, it returns the appropriate boolean based on the
success of the db modification.

getJSON (renamed from generateJSON) now builds an appropriate JSON
object depending on the given eventType. It uses functions phoneGets,
getEvents, getPies, and checkLogin.

phoneGets deals with phones' GET requests, returning a list of
JSON-formatted objects containing event information. It uses getEvents
to get all the appropriate events (depending on the event type and
option).

getPies returns a list of raspberry ids connected to a
specified user.checkLogin just returns simple json, which depends on whether
the login details are correct or not.

Also added functions connectUserToRaspberry (self-explanatory), and
checkUser, which checks if a username/email already exists in the
database and returns a boolean.